+++ common/dynamicconfig/constants.go
@@ -65,14 +65,14 @@ var Keys = map[Key]string{
 	AdminMatchingNamespaceTaskqueueToPartitionDispatchRate: "admin.matchingNamespaceTaskqueueToPartitionDispatchRate",
 
 	// system settings
-	StandardVisibilityPersistenceMaxReadQPS:  "system.standardVisibilityPersistenceMaxReadQPS",
-	StandardVisibilityPersistenceMaxWriteQPS: "system.standardVisibilityPersistenceMaxWriteQPS",
-	AdvancedVisibilityPersistenceMaxReadQPS:  "system.advancedVisibilityPersistenceMaxReadQPS",
-	AdvancedVisibilityPersistenceMaxWriteQPS: "system.advancedVisibilityPersistenceMaxWriteQPS",
-	AdvancedVisibilityWritingMode:            "system.advancedVisibilityWritingMode",
-	EnableWriteToSecondaryVisibility:         "system.enableWriteToSecondaryVisibility",
-	EnableReadVisibilityFromES:               "system.enableReadVisibilityFromES",
-	EnableReadFromSecondaryVisibility:        "system.enableReadFromSecondaryVisibility",
+	StandardVisibilityPersistenceMaxReadQPS:   "system.standardVisibilityPersistenceMaxReadQPS",
+	StandardVisibilityPersistenceMaxWriteQPS:  "system.standardVisibilityPersistenceMaxWriteQPS",
+	AdvancedVisibilityPersistenceMaxReadQPS:   "system.advancedVisibilityPersistenceMaxReadQPS",
+	AdvancedVisibilityPersistenceMaxWriteQPS:  "system.advancedVisibilityPersistenceMaxWriteQPS",
+	AdvancedVisibilityWritingMode:             "system.advancedVisibilityWritingMode",
+	EnableWriteToSecondaryAdvancedVisibility:  "system.enableWriteToSecondaryAdvancedVisibility",
+	EnableReadVisibilityFromES:                "system.enableReadVisibilityFromES",
+	EnableReadFromSecondaryAdvancedVisibility: "system.enableReadFromSecondaryAdvancedVisibility",
 
 	HistoryArchivalState:                   "system.historyArchivalState",
 	EnableReadFromHistoryArchival:          "system.enableReadFromHistoryArchival",
@@ -385,12 +385,12 @@ const (
 	AdvancedVisibilityPersistenceMaxWriteQPS
 	// AdvancedVisibilityWritingMode is key for how to write to advanced visibility
 	AdvancedVisibilityWritingMode
-	// EnableWriteToSecondaryVisibility is the config to enable write to secondary visibility for elastic search
-	EnableWriteToSecondaryVisibility
+	// EnableWriteToSecondaryAdvancedVisibility is the config to enable write to secondary visibility for elastic search
+	EnableWriteToSecondaryAdvancedVisibility
 	// EnableReadVisibilityFromES is key for enable read from elastic search
 	EnableReadVisibilityFromES
-	// EnableReadFromSecondaryVisibility is the config to enable read from secondary elastic search
-	EnableReadFromSecondaryVisibility
+	// EnableReadFromSecondaryAdvancedVisibility is the config to enable read from secondary elastic search
+	EnableReadFromSecondaryAdvancedVisibility
 	// DisableListVisibilityByFilter is config to disable list open/close workflow using filter
 	DisableListVisibilityByFilter
