+++ common/persistence/visibility/visibility_manager_dual.go
@@ -25,89 +25,36 @@
 package visibility
 
 import (
-	"fmt"
-
-	"go.temporal.io/api/serviceerror"
-	"go.temporal.io/server/common/dynamicconfig"
-	"go.temporal.io/server/common/namespace"
 	"go.temporal.io/server/common/persistence/visibility/manager"
 )
 
 type (
 	visibilityManagerDual struct {
-		stdVisibilityManager manager.VisibilityManager
-		advVisibilityManager manager.VisibilityManager
-		managerSelector      managerSelector
-	}
-
-	managerSelector interface {
-		readManager(namespace namespace.Name) manager.VisibilityManager
-		writeManagers() ([]manager.VisibilityManager, error)
-	}
-
-	sqlToESManagerSelector struct {
-		enableAdvancedVisibilityRead  dynamicconfig.BoolPropertyFnWithNamespaceFilter
-		advancedVisibilityWritingMode dynamicconfig.StringPropertyFn
-		stdVisibilityManager          manager.VisibilityManager
-		advVisibilityManager          manager.VisibilityManager
-	}
-
-	esManagerSelector struct {
-		enableReadFromSecondaryVisibility dynamicconfig.BoolPropertyFnWithNamespaceFilter
-		enableWriteToSecondaryVisibility  dynamicconfig.BoolPropertyFn
-		visibilityManager                 manager.VisibilityManager
-		secondaryVisibilityManager        manager.VisibilityManager
+		visibilityManager          manager.VisibilityManager
+		secondaryVisibilityManager manager.VisibilityManager
+		managerSelector            managerSelector
 	}
 )
 
 var _ manager.VisibilityManager = (*visibilityManagerDual)(nil)
-var _ managerSelector = (*sqlToESManagerSelector)(nil)
-var _ managerSelector = (*esManagerSelector)(nil)
 
-// NewVisibilityManagerDual create a visibility manager that operate on DB or Elasticsearch based on dynamic config.
+// NewVisibilityManagerDual create a visibility manager that operate on multiple manager
+// implementations based on dynamic config.
 func NewVisibilityManagerDual(
-	stdVisibilityManager manager.VisibilityManager,
-	advVisibilityManager manager.VisibilityManager,
+	visibilityManager manager.VisibilityManager,
+	secondaryVisibilityManager manager.VisibilityManager,
 	managerSelector managerSelector,
 ) *visibilityManagerDual {
 	return &visibilityManagerDual{
-		stdVisibilityManager: stdVisibilityManager,
-		advVisibilityManager: advVisibilityManager,
-		managerSelector:      managerSelector,
-	}
-}
-
-func NewSQLToESManagerSelector(
-	stdVisibilityManager manager.VisibilityManager,
-	advVisibilityManager manager.VisibilityManager,
-	enableAdvancedVisibilityRead dynamicconfig.BoolPropertyFnWithNamespaceFilter,
-	advancedVisibilityWritingMode dynamicconfig.StringPropertyFn,
-) *sqlToESManagerSelector {
-	return &sqlToESManagerSelector{
-		stdVisibilityManager:          stdVisibilityManager,
-		advVisibilityManager:          advVisibilityManager,
-		enableAdvancedVisibilityRead:  enableAdvancedVisibilityRead,
-		advancedVisibilityWritingMode: advancedVisibilityWritingMode,
-	}
-}
-
-func NewESManagerSelector(
-	visibilityManager manager.VisibilityManager,
-	secondaryVisibilityManager manager.VisibilityManager,
-	enableReadFromSecondaryVisibility dynamicconfig.BoolPropertyFnWithNamespaceFilter,
-	enableWriteToSecondaryVisibility dynamicconfig.BoolPropertyFn,
-) *esManagerSelector {
-	return &esManagerSelector{
-		visibilityManager:                 visibilityManager,
-		secondaryVisibilityManager:        secondaryVisibilityManager,
-		enableReadFromSecondaryVisibility: enableReadFromSecondaryVisibility,
-		enableWriteToSecondaryVisibility:  enableWriteToSecondaryVisibility,
+		visibilityManager:          visibilityManager,
+		secondaryVisibilityManager: secondaryVisibilityManager,
+		managerSelector:            managerSelector,
 	}
 }
 
 func (v *visibilityManagerDual) Close() {
-	v.stdVisibilityManager.Close()
-	v.advVisibilityManager.Close()
+	v.visibilityManager.Close()
+	v.secondaryVisibilityManager.Close()
 }
 
 func (v *visibilityManagerDual) GetName() string {
@@ -208,40 +155,4 @@ func (v *visibilityManagerDual) ScanWorkflowExecutions(request *manager.ListWork
 
 func (v *visibilityManagerDual) CountWorkflowExecutions(request *manager.CountWorkflowExecutionsRequest) (*manager.CountWorkflowExecutionsResponse, error) {
 	return v.managerSelector.readManager(request.Namespace).CountWorkflowExecutions(request)
-}
-
-func (v *sqlToESManagerSelector) writeManagers() ([]manager.VisibilityManager, error) {
-	switch v.advancedVisibilityWritingMode() {
-	case AdvancedVisibilityWritingModeOff:
-		return []manager.VisibilityManager{v.stdVisibilityManager}, nil
-	case AdvancedVisibilityWritingModeOn:
-		return []manager.VisibilityManager{v.advVisibilityManager}, nil
-	case AdvancedVisibilityWritingModeDual:
-		return []manager.VisibilityManager{v.stdVisibilityManager, v.advVisibilityManager}, nil
-	default:
-		return nil, serviceerror.NewInternal(fmt.Sprintf("Unknown advanced visibility writing mode: %s", v.advancedVisibilityWritingMode()))
-	}
-}
-
-func (v *sqlToESManagerSelector) readManager(namespace namespace.Name) manager.VisibilityManager {
-	if v.enableAdvancedVisibilityRead(namespace.String()) {
-		return v.advVisibilityManager
-	}
-	return v.stdVisibilityManager
-}
-
-func (v *esManagerSelector) writeManagers() ([]manager.VisibilityManager, error) {
-	managers := []manager.VisibilityManager{v.visibilityManager}
-	if v.enableWriteToSecondaryVisibility() {
-		managers = append(managers, v.secondaryVisibilityManager)
-	}
-
-	return managers, nil
-}
-
-func (v *esManagerSelector) readManager(namespace namespace.Name) manager.VisibilityManager {
-	if v.enableReadFromSecondaryVisibility(namespace.String()) {
-		return v.secondaryVisibilityManager
-	}
-	return v.visibilityManager
 }
