--- schema/elasticsearch/visibility/versioned/v2/index_template_v6.json
@@ -1,71 +1,0 @@
-{
-  "order": 0,
-  "index_patterns": [
-    "temporal_visibility_v2*"
-  ],
-  "settings": {
-    "index": {
-      "number_of_shards": "1",
-      "number_of_replicas": "0",
-      "auto_expand_replicas": "0-2",
-      "sort.field": [ "CloseTime", "StartTime", "RunId" ],
-      "sort.order": [ "desc", "desc", "desc" ]
-    }
-  },
-  "mappings": {
-    "_doc": {
-      "dynamic": "false",
-      "properties": {
-        "NamespaceId": {
-          "type": "keyword"
-        },
-        "WorkflowId": {
-          "type": "keyword"
-        },
-        "RunId": {
-          "type": "keyword"
-        },
-        "WorkflowType": {
-          "type": "keyword"
-        },
-        "StartTime": {
-          "type": "date"
-        },
-        "ExecutionTime": {
-          "type": "date"
-        },
-        "CloseTime": {
-          "type": "date"
-        },
-        "ExecutionDuration": {
-          "type": "long"
-        },
-        "ExecutionStatus": {
-          "type": "keyword"
-        },
-        "TaskQueue": {
-          "type": "keyword"
-        },
-        "TemporalChangeVersion": {
-          "type": "keyword"
-        },
-        "BatcherNamespace": {
-          "type": "keyword"
-        },
-        "BatcherUser": {
-          "type": "keyword"
-        },
-        "BinaryChecksums": {
-          "type": "keyword"
-        },
-        "HistoryLength": {
-          "type": "long"
-        },
-        "StateTransitionCount": {
-          "type": "long"
-        }
-      }
-    }
-  },
-  "aliases": {}
-}
