+++ service/frontend/fx.go
@@ -302,7 +302,8 @@ func VisibilityManagerProvider(
 	return visibility.NewManager(
 		params.PersistenceConfig,
 		persistenceServiceResolver,
-		esConfig,
+		esConfig.GetVisibilityIndex(),
+		esConfig.GetSecondaryVisibilityIndex(),
 		esClient,
 		nil, // frontend visibility never write
 		saProvider,
@@ -313,7 +314,7 @@ func VisibilityManagerProvider(
 		serviceConfig.AdvancedVisibilityPersistenceMaxWriteQPS,
 		serviceConfig.EnableReadVisibilityFromES,
 		dynamicconfig.GetStringPropertyFn(visibility.AdvancedVisibilityWritingModeOff), // frontend visibility never write
-		serviceConfig.EnableReadFromSecondaryVisibility,
+		serviceConfig.EnableReadFromSecondaryAdvancedVisibility,
 		dynamicconfig.GetBoolPropertyFn(false), // frontend visibility never write
 		params.MetricsClient,
 		logger,
