+++ service/history/configs/config.go
@@ -50,7 +50,7 @@ type Config struct {
 	AdvancedVisibilityPersistenceMaxReadQPS  dynamicconfig.IntPropertyFn
 	AdvancedVisibilityPersistenceMaxWriteQPS dynamicconfig.IntPropertyFn
 	AdvancedVisibilityWritingMode            dynamicconfig.StringPropertyFn
-	EnableWriteToSecondaryVisibility         dynamicconfig.BoolPropertyFn
+	EnableWriteToSecondaryAdvancedVisibility dynamicconfig.BoolPropertyFn
 
 	EmitShardDiffLog      dynamicconfig.BoolPropertyFn
 	MaxAutoResetPoints    dynamicconfig.IntPropertyFnWithNamespaceFilter
@@ -294,7 +294,7 @@ func NewConfig(dc *dynamicconfig.Collection, numberOfShards int32, isAdvancedVis
 		AdvancedVisibilityPersistenceMaxReadQPS:  dc.GetIntProperty(dynamicconfig.AdvancedVisibilityPersistenceMaxReadQPS, 9000),
 		AdvancedVisibilityPersistenceMaxWriteQPS: dc.GetIntProperty(dynamicconfig.AdvancedVisibilityPersistenceMaxWriteQPS, 9000),
 		AdvancedVisibilityWritingMode:            dc.GetStringProperty(dynamicconfig.AdvancedVisibilityWritingMode, visibility.DefaultAdvancedVisibilityWritingMode(isAdvancedVisibilityConfigExist)),
-		EnableWriteToSecondaryVisibility:         dc.GetBoolProperty(dynamicconfig.EnableWriteToSecondaryVisibility, false),
+		EnableWriteToSecondaryAdvancedVisibility: dc.GetBoolProperty(dynamicconfig.EnableWriteToSecondaryAdvancedVisibility, false),
 
 		EmitShardDiffLog:                     dc.GetBoolProperty(dynamicconfig.EmitShardDiffLog, false),
 		HistoryCacheInitialSize:              dc.GetIntProperty(dynamicconfig.HistoryCacheInitialSize, 128),
